Revision of wave-atmosphere coupling
See merge request icon/icon-nwp!1669

The coupling of the wave model with the atmosphere, specifically the usage of the z0 (roughness length) field passed from the wave model to the atmosphere, is revised. 

Currently, z0_waves simply replaces the roughness length usually diagnosed in turbtran, disregarding the molecular contribution to z0. This strongly increases the surface fluxes over oceans and induces an unacceptably large warm bias throughout the troposphere.

In the revised implementation, z0_waves only replaces the Charnock-based component of the turbtran z0, and only if z0_waves is at least 0.1 mm. This leads to almost neutral scores and provides a reasonable starting point for future tuning work.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>