Aerosol-cloud interaction for cloudice2mom using ART dust for ice nucleation
See merge request icon/icon-nwp!1593

<!--
ICON

---------------------------------------------------------------
Copyright (C) 2004-2024, DWD, MPI-M, DKRZ, KIT, ETH, MeteoSwiss
Contact information: icon-model.org

See AUTHORS.TXT for a list of authors
See LICENSES/ for license information
SPDX-License-Identifier: CC0-1.0
---------------------------------------------------------------
-->
## Feature Description

Coupling with ICON-ART dust as ice nucleating particles depending on icpl_aero_ice. Here two options are provided, one using a full coupling with aerosol size information from ART (icpl_aero_ice=4) and a simplified coupling with fixed modal diameters (icpl_aero_ice=3). The latter is currently recommended on DWD's NEC. Other available ice nucleation options are icpl_aero_ice=0 for the Cooper formula and icpl_aero_ice=2 for the INAS ice nucleation scheme with an internally prescribed dust profile. The DeMott scheme (icpl_aero_ice=1) is currently not supported for the two-moment cloud ice scheme. These options are described in the namelist documentation.

The subroutines to calculate the aerosol properties needed for the INAS approach, namely the total number and mean surface area of dust particles, are aerosol_prepare_inas_dust in mo_cpl_aerosol_microphys in the NWP repository for icpl_aero_ice=3 and art_prepare_dust_inas in mo_art_prepare_aerosol in the ART repository for icpl_aero_ice=4. This has been approved by KIT (Ali Hoshyaripour) and DWD (Günther Zängl). The subroutine aerosol_prepare_inas_dust can later be extended to include other (external) prognostic aerosol models.

This is the 3rd part of the former !1479. The 2nd part was !1581.

With the MR we also wanted to fix an issue introduced by !1612, which broke the 1mom microphysics granule test, when it was merged on 9 Dec 2024. Unfortunately, Daniel Rieger rejected to merge this MR as a Gatekeeper before the Christmas holidays. Due to this delay, Jonas Jucker fixed the granule test in a separate MR !1633 on 23 Dec, but did unfortunately not reduce the list of dependencies as it was planned to do with the current MR (and as it has been discussed with Pavel Khain). Hence, here we still do this minor cleanup to the 2mom and SBM microphysics, which reduces the dependencies of the 1mom granule.

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>