[nwp] Wave: Fix usage of bathymetry versus water depth field
See merge request icon/icon-nwp!1667

In the wave model we need to distinguish the (time constant) bathymetric height `bathymetry_[c/e]` from the (potentially time varying) water depth `depth_[c/e]`. This branch is an attempt to fix the usage of bathymetry versus depth throughout the wave code. No change in results is expected for the available testcases, as no time-varying sea level height information is read so far. This changeset should be regarded as preparatory work in order to deal with time-dependent water depth in the future. 


#### Cleanup

The memory footprint is reduced slightly, by removing the obsolete 2D fields `sea_ice_e` and `sea_level_e`. 


#### Bug fixes

* [x] So far, an additional limitation of `bathymetry_e` was required in order to avoid a division by zero in the subroutine `wave_number_e`. This has been fixed by ensuring that the cell to edge interpolation of `bathymetry_c` takes care of the boundary edges as well.
* [x] Fix coupled atm-wave experiment `run_ICON_28_R2B6_IAU_atmo_waves_coupled`. Due to previous modifications regarding output variable names https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1663, only a small subset of fields was read from the restart file for the wave model. Missing or unrecognized fields have simply been ignored without throwing an error message. 
This issue has been solved by generating a new restart file with proper variable names. In addition the experiment has been modified such that wave performs a strict restart rather than an initialization from restart. These modes differ with regards to the behavior when one or more fields are missing in the restart file. In case of a strict restart, the model will abort if one or more fields are missing in the restart file.


@u241211

Approved-by: Mikhail Dobrynin <mikhail.dobrynin@dwd.de>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>