[mpim] Fix ncontained counter for add_ref usage
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/614

This MR hacks the 'ncontained' counter such that tracer allocation by, e.g.,
ComIn plays nicely with field or slice aliases. This hack does not solve the
fundamental problem that 'add_ref' is used to create "true" variable or an alias
name for a slice depending on a context. It is still possible to create gaps and
'ncontained' does neither count the number of references nor the number of
aliases. It is now just a counter for the maximum index used by the slice.

Co-authored-by: Nils-Arne Dreier <dreier@dkrz.de>
Approved-by: Nils-Arne Dreier <dreier@dkrz.de>
Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>