[mpim] Fix for runscript generation of bubble test cases with NAG on levante
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/658

When generating a runscript for one of the `exp.aes_bubble*` experiments the
default on levante is currently that they run on the compute partition with
128 procs. This leads to a crash with the NAG compiler during domain
decomposition.

This MR fixes this by changing `run/create_target_header` to use the shared
partition and 16 procs for the bubble experiments. This has not been a problem
in buildbot because buildbot already used shared/16 procs in the generation of
the test lists.

Closes #83

Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>