Check Memory consumption on NEC
See merge request icon/icon-nwp!1399

## Feature Description
This MR analyzes the memory report from the NEC `mpirun` to make sure that the memory consumption on the VE does not increase significantly without notice. This MR implements an upper and lower mem check. The lower mem check is to make sure that significant optimization is detected so that the limits can be adjusted thoroughly.

The memory monitoring of the NEC is based on sampling. Therefore, the reporting is stochastic. To reduce the number of false alarms, the test run is repeated once if a threshold is exceeded.

The limits are based on 172, 533, and 160 runs of experiments 03, 09, and 14, respectively. I choose the minimum of all runs of each experiment as lower limit and  216/215 times the minimum as the upper limit. (The factor is based on the observation that the memory footprint of ICON is approximately on the order of 215 3D fields)

(This MR also fixes the OMP tests on the hybrid builder after renaming of the builders)

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>