Corrections (mainly) of extended options related to the turbulence scheme: Part I
See merge request icon/icon-nwp!1671

The turbulence scheme, represented by the modules 'turb_[diffusion|tansfer|utlities|vertdiff|data]' already contains a couple of extensions or alternative modes, which are going to be consolidated or reduced.
With this MR, the matter of transmitting additional shear-forcing at the lowest ordinary atmospheric half-level "P" (k=ke) - as calculated by 'turb_diffusion' - towards the "0"-level (k=ke1) - treated by 'turb_transfer' - is being considerably revised.

This is intrinsically connected with the additional shear generation by Non-Turbulent-Circulations (NTCs) - as described through the STIC-concept - as well as to the treatment of Lower Limits of Diffusion Coefficients (LLDCs), and it also matters the representation of the laminar limit, the prevention of singulatities in the solution of the Turbulence Model, the initialization procedue, the calculation of transfer-layer profile-functions and the generation of sea-surface roughness.

The foreseen measures are related mainly to the following parameters:
'rsur_shear', 'imode_suradap', 'imode_nsf_wind', 'imode_trancnf', 'imode_tkemini', 'imode_lamdiff'
'stbsecu', 'imode_stbcorr->imode_stbcalc'

In the course of this revision and correction, also some simplifications as well as reductions of selector-options are being applied. Further a couple of comments are goning to be adapted for a better understanding.  

After this first correction project 'mrcor1' some pure reductions of no longer required options are planned, starting with 'mrred1'.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>