Add accumulated and max/min diagnostic fields to nest start interpolation
See merge request icon/icon-nwp!1682

Accumulated/averaged and max/min diagnostic fields are added to the nest start interpolation in order to obtain correct output fields if a nested domain is started during runtime (as for ICON-D05). Analogous to `interpol_phys_grf`, a distinction is now made between positive definite fields and others, which means that for the former group, a gradient limiter is applied during the parent-to-child interpolation. This slightly changes the forecast results in nested domains started during runtime because the interpolated precipitation rates are used in the first TERRA call after the nest start.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>