[mpim] Bugfix for the generation of index lists on GPUs (LAM and nested)
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/592

For LAM or nested simulations, boundary cells are located at the beginning of the arrays.
Hence, loops over all computation cells often start at element ics, with ics > 1. In the
subroutines 'generate_index_list' and 'generate_index_list_batched', this causes problems
on GPUs as the GPU-specific routine receives a pointer to the first element of the condition
array. The bug is solved by fixing the cuda and hip backend functions.

Closes https://gitlab.dkrz.de/icon/icon-mpim/-/issues/166

Co-authored-by: Dmitry Alexeev <dalexeev@nvidia.com>
Approved-by: Moritz Waldmann <moritz.waldmann@dwd.de>
Approved-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>