Reducing the complexity of the operational turbulence scheme for NWP - Part 2:  Removal of some undesired options
See merge request icon/icon-nwp!1694

The operational turbulence scheme for NWP (based on the modules turb_[diffusion|transfer|vertdiff|utilities|data]) can be controlled by a couple of parameters, switches and selectors.
Some of them are related to obsolete options or belong to some alternative treatment, which appears to be dispensable. By this MR, some options are removed, which don't show any positive impact so far, can be dropped without violation of consistency, and are neither a focus of the current development nor do represent a selection of considered extensions or assumptions. 
These deleted options refer to the sitches 'lprfcor', lnonloc', 'lstbsmot, 'ltndsmot' and to the case "imode_trancnf=1".

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>