[mpim] ICON-Land: Updates for the uniform-scale soil hydrology
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/532

Features new formulations for the treatment of lateral fluxes at and below the
surface in case the soil hydrology is used with uniform scale parametrizations.

This includes updates to the formulations for infiltration / runoff generation
and bottom layer drainage. Additionally, a parametrization for lateral drainage
from all soil layers above the bedrock is included. As a general feature, the
generated surface and subsurface runoff is not assumed to immediately contribute
to the streamflow, but is initially stored in an intermediary reservoir. From
this reservoir the water moves to the streamflow with a fixed time lag and
drains at the bottom of the soil column but can also move back into the soil in
case the latter dries out.

These changes only affect simulations that are conducted using the uniform scale
hydrology setup. Results of the the standard setup (semi-distributed case) remain
unchanged.

For details check jsbach MR https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/202

Co-authored-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>
Co-authored-by: Tobias Stacke <tobias.stacke@mpimet.mpg.de>
Co-authored-by: Philipp De-Vrese <philipp.de-vrese@mpimet.mpg.de>
Merged-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>