[mpim] Fix create_target_header for aes_bubble tests
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/699

In https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/658, a section in "create_target_header"
was added to run "*aes_bubble*" experiments on levante CPU on the shared partition using 16
processors because using the compute partition with the default 128 processors doesn't work.

It turned out that the line "use_mpi_procs_pernode=16" doesn't work if OpenMP is enabled because,
by fixing "mpi_procs_pernode", the number of OpenMP threads is not considered.

This MR fixes this by deleting this line ... it is sufficient to set "use_tasks=16".

Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>