[nwp] Implementation of new CAMS climatology  (49R2, aerosol mixing ratios instead of layer masses)
See merge request icon/icon-nwp!1632

The format of the CAMS aerosol climatologies has changed, and aerosol is now provided as mixing ratio (kg/kg) instead of layer integrated mass. There is also a new climatology (49R2) which corrects some unrealistic artefacts that were present in the previous climatology version (43R3), e.g. too much dust in the stratosphere. This branch changes how the climatology is read in and interpolated in the vertical to the ICON vertical grid.
This branch is NOT backwards-compatible and no longer contains the option to read in the older 43R3 CAMS climatology.

Documentation on the new CAMS implementation has been provided, and can be found here:
https://gitlab.dkrz.de/icon-model-org/website/-/merge_requests/52

Approved-by: Harel Muskatel <muskatelh@ims.gov.il>
Merged-by: Daniel Rieger <377-m300368@users.noreply.gitlab.dkrz.de>