Reducing the complexity of the operational turbulence scheme for NWP - Part 3: Removal of further obsolete options
See merge request icon/icon-nwp!1697

The operational turbulence scheme for NWP (based on the modules turb\_\[diffusion|transfer|vertdiff|utilities|data\]) can be controlled by a couple of parameters, switches and selectors. Some of them are related to obsolete options or belong to some alternative treatment, which appears to be dispensable.
As an amendment of "Part 2" (branch 'mrred2', MR !1194), further options are removed by this MR now, which also belong to the characteristics of dispensable options, as mentioned in the description of "Part 2".
These are "imode_2m_diag=(-)1", "imode_2m_diag<0" and "imode_calcirc=1" (at which 'imode_calcirc' and 'imode_2m_diag' vanishes), since all these options mainly express versions of the former COSMO-code, which is no longer being updated by ICON-code.
Moreover, the option "imode_qvsatur=1" is removed (at which 'imode_qvsatur' vanishes as well), and the no longer required flexibility related to the logical parameters 'l\[uv\]m_dif\\|lscadif' is removed now in SUB 'vertdiff', as well as the parameter 'lstfnct' is no longer present in the parameter lists of turbulence routines.
Finally, the parameter 'lexpcor' lacks in the 'turbdiff'-NAMELIST, until some related issues are resolved.

For the option "ltmpcor=T", which activates the calculation of heating by TKE dissipation, a consistency fix has been implemented that skips the respective approximate calculations done for the SSO and GWD schemes in the NWP interface, in order to prevent respective double-counting.
Furthermore, a security measure has been implemented for option "ltmpcor=T" to ensure numerical stability, and the selector 'imode_tkemini' is a NAMELIST parameter now.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>