[mpim] Bugfix for array initialization in coupled simulations using GPUs
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/711

This MR fixes the initialization of `get_buffer` array in 
`coupling/mo_aes_ocean_coupling.f90` in simulations using GPUs. Currently, 
the array is initialized on the GPU by calling `init` with `lacc=.TRUE`.
However, the array is then filled by YAC on the CPU-side and pushed to the
GPU with ACC UPDATE, making the initialization of GPU array useless.
Because YAC doesn't touch masked or halo cells, possible invalid values in
the uninitialized CPU array can remain and cause FPEs. The MR changes 
`lacc` to `.FALSE.` in initialization calls so that the array is properly
cleared.

Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>