improve coupled-model water conservation
See merge request icon/icon-nwp!1465

To ensure that the coupled model with ocean and hydrological discharge (HD) components has a closed water balance to the best extent possible, we add TERRA's own water nonconservation diagnostic to the subsurface runoff that gets passed to the HD model. This formally closes the budget up to rounding errors.  Note that lake water balance is already closed by setting the surface runoff to P-E, which gets passed to the HD model as well. The MR also adds detailed diagnostic output to the HD coupling for message level 20.

This work has been done together with Roland Wirth @b381588.

Approved-by: Roland Wirth <roland.wirth@dwd.de>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>