Update NWP test 03 in order to execute the code running in ICON-D05
See merge request icon/icon-nwp!1722

### Short description

The NWP buildbot test 03 is updated in order to execute the code running in ICON-D05 that has recently become operational. This involves switching from 2-moment microphysics to the 1-moment graupel scheme, which is uncritical because dedicated buildbot tests for the RUC configuration have been added some time ago. To save resources, the parameterizations running at 500 m in the operational configuration are executed in the 1-km nest in the buildbot test.

The input data are the same as for test 08, implying that the old input data for test 03 can be removed (or read-protected) from the DWD pool directory after pushing the update to the main master.

The resource allocation remains unchanged at 6 VEs because 5 would be insufficient. The upper and lower limits for the memory allocation check are adjusted.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>