Option for prognostic anthropogenic heat flux based on time-filtered T2M
See merge request icon/icon-nwp!1689

An option for a time-dependent specification of the anthropogenic heat flux (AHF) in introduced (`itype_ahf=3`). It differs from the default option (`itype_ahf=2`) in taking a time-filtered T2M rather a climatological T2M as input. This to some degree allows for accounting for the fact that the AHF is larger during wintertime cold spells and summertime heat waves than under average conditions.

Approved-by: Jan-Peter Schulz <jan-peter.schulz@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>