New NWP physics options for tuning and improved physical consistency
See merge request icon/icon-nwp!1719

### Short description

A new option (`itype_dissip_heat=2`) is introduced for taking into account dissipative heating from turbulent momentum dissipation in the turbulence interface, complementing the default (dissipative heating from SSO and GWD). When dissipative heating is computed within the turbulence scheme, the respective computations in the interface are turned off.

In addition, an improved option for ozone-tropopause coupling (`icpl_o3_tp=2`) is introduced that avoids adding excessive amounts of ozone for low tropopauses. Moreover, the option `itune_o3=3`, which has become obsolete with the update of the greenhouse gas concentrations made in 2024, is modified and supposed to be combined with `icpl_o3_tp=2`.

To allow adjusting the temperature and radiation biases of the global EPS mean to those of the deterministic forecasts, the parameters `shift_ratsea` and `shift_boxliq_asy` are introduced into the ensemble namelist. Although this could be done manually in the namelists as well, this way is believed to be less error-prone because the need for these shift parameters is not a matter of resolution, which otherwise might cause confusion for the ART ensemble operated at DWD.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>