[nwp] Bugfix in homogeneous ice nucleation of cloudice2mom (gscp3)
See merge request icon/icon-nwp!1720

## Feature Description

Bugfix in homogeneous ice nucleation of cloudice2mom. There was a confusion between specific mass and mean mass in the parameterization of homogeneous ice nucleation. The change in the results is minor due to the limiters which have partly masked the bug. The bugfix leads to an increase of the reflected shortwave radiation at TOA by 0.2-0.3 W/m2 in a R03B06 simulation of ICON-ART. This has been tested for Feb-March-April 2024 and the agreement with CERES SYN1DEG is actually better after the bugfix.

Approved-by: Carmen Koehler <carmen.koehler@dwd.de>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>