[nwp] Bug fix for reading first guess of t_seasfc
See merge request icon/icon-nwp!1743

For correct first-guess cycling of `t_seasfc`, it is necessary to reset `cdi_missval`s to zero after reading this field. Due to a mistake made while cleaning up the code for !1469, this reset is not executed when analysis and first guess are read at the same time. This is corrected now.

@b381078

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>