[nwp] Bug fix for rime formation term for interception storage
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1740

### Short description

The rime formation term in TERRA, which is one of the source terms for the interception storage, erroneously uses the upper soil temperature rather than the skin temperature in the absence of snow. This underestimates the formation of rime, leading to excessive fog formation in some situations. This error is now fixed.

In addition, evaporation from the interception is retuned in order to be more active at temperatures below and slightly above freezing. This turned out to improve the diurnal cycle of 2-m humidity.

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Approved-by: Roland Wirth <roland.wirth@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>