[nwp] Using YAC for reading and interpolating ozone and Kinne aerosol data also in icon-nwp
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1716

## Feature Description
One advantage of using YAC for reading ozone and Kinne aerosol is both ozone and aerosol (AOD etc.) data can be provided independently of the used ICON grid. The remapping to the ICON grid as well as the temporal interpolation is done by YAC. This is an extension to use the existing functionality for ecrad as well. 
 
The handling remains the same as for rrtmgp, controlling the YAC via coupling_mode_nml with coupled_to_aero = True and coupled_to_o3 = True, as well as via coupling.yaml. Here distance weighted interpolation is used for the interpolation to the ICON grid. 
 
The functionality can be used for irad_o3 = 5 (transient ozone for the years 1850 - 2014) and irad_aero = 13 (1850 Kinne background + transient for the years 1850-2100). 
 

Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Co-authored-by: Rene Redler <rene.redler@mpimet.mpg.de>
Co-authored-by: Andrea Schneidereit <m300590@levante3.lvt.dkrz.de>
Co-authored-by: Ralf Mueller <ralf.mueller@dkrz.de>
Co-authored-by: Moritz Hanke <hanke@dkrz.de>
Approved-by: Roland Wirth <roland.wirth@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>