[nwp] Regime-dependent FSD parameterization
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1726

This branch introduces the option to use a regime-dependent FSD parameterization instead of a globally fixed value of 1. This option is set to .false. by default.

At this point in time, the parameterization shows no particular benefit in the global ICON forecast verification. However, it reduces TOA radiation RMSE and is therefore of interest for ICON ESM/seamless/XPP configurations.

Co-authored-by: Guenther Zaengl <Guenther.Zaengl@dwd.de>
Co-authored-by: Sophia Schaefer <sophia.schaefer@dwd.de>
Co-authored-by: Martin Köhler <martin.koehler@dwd.de>
Approved-by: Mareike Burba <mareike.burba@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>