[nwp] Adding latex like syntax support for doc/www
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1754

The sphinx documentation describes two ways to render latex-like equation syntax. One is to invoke latex to generate a png, the second is based on JavaScript which can be invoked externally or internally. The JavaScript version is prefered by DKRZ since it is cleaner and easier. The internal invocation is better wrt. data protection.

This MR adds the necessary script from MathJax and the necessary adaptions to the CI-pipelines.

Approved-by: Jan Frederik Engels <engels+git@dkrz.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>