[nwp] Fix echotop and echotopinm diagnostics for clouds reaching model top
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1750

## Feature Description
This MR is a bug fix for the ECHOTOP and ECHOTOPinM diagnostics, which determine the lowest pressure respectively largest height where given 
radar reflectivity thresholds are exceeded.

If in extreme cases
- intense clouds reach the model top and
- exceed the echotop threshold in all leves and 
- reflectivity is larger at model top than one layer below

the diagnostic produced erroneous values, which in case of echotopinm are on the order of 1E+303, which are a valid double precision values (so ICON does not crash) but cannot be written to grib files lead to a crash of eccodes.

This has been fixed in a way that in the above pathological case the echotop's are set to the model top values of pressure and height.

**Otherwise, no change of results.**

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Approved-by: Klaus Stephan <klaus.stephan@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>