[nwp] Avoid calling SIZE within GPU code (especially for loop bounds)
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1745

## Feature Description
This MR includes safety measures to avoid weird bugs based on previous investigations.
It removes calls to SIZE() when specifying loop bounds.

The measures to avoid weird bugs based on the investigations for [this issue](https://gitlab.dkrz.de/icon/wiki/-/wikis/GPU-development/Solved:-Get-rid-of-mysterious-variables). We moved calls to Fortran's `SIZE()` outside GPU code when possible and especially if it was called to set loop bounds "in place".

Note that these changes are not exhaustive, but they simply come from an (unsuccessful) attempt to fix the bug described in MR !1742.

Co-authored-by: Annika Lauber <annika.lauber@c2sm.ethz.ch>
Co-authored-by: Mikael Stellio <mikael.stellio@meteoswiss.ch>
Approved-by: Annika Lauber <annika.lauber@c2sm.ethz.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>