[nwp] Fix: Enabling "c_diff=0" in case of "lcircterm=T" in NWP turbulence.
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1768

### Short description

In the course of prior modifications, choosing no vertical diffusion of TKE (c_diff=0.0) for NWP-turbulence causes a division-by-zero error, if the additional TKE-source by near-surface thermals (the "circulation term") is enabled (pat_len>0.0). This problem occurs, because the "circulation term" is treated as the convergence of an additional non-gradient TKE-flux.
With this fix, "c_diff=0.0" works in any configuration.

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>