[mpim] Routines to read in First Guess and Analysis Files in the Ocean
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/511

Implementation of Routines to read first guess files and analysis files in grb-format.
For this the cdi-library is used. Implementation strongly follows the initicon-implementation in atm-code. Additionally, a timeshift analogous to timeshift dt_shift is implemented in the ocean,
so that a timeshift fitting the atmospheric timeshift for IAU is possible.


Co-authored-by: Daniel Krueger <Daniel.Krueger@dwd.de>
Co-authored-by: Nora Schenk <nora.schenk@dwd.de>

Merged-by: Helmuth Haak <helmuth.haak@mpimet.mpg.de>