[nwp] Stochastic pattern generator based on spherical harmonics
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1708

## Feature Description
<!--Describe your feature and important technical aspects-->

This MR provides a new module which implements the stochastic pattern generator (SPG) of Palmer et al. (2009, their Appendix 8.1). This pattern generator uses spherical harmonics for the global ICON and Fourier representation for limited-area. At ECWMF this pattern generator has been used for SPPT for many years and the same stochastic pattern generator is now used for the SPP as described in Ollinaho et al. (2017). The limited-area version has been used, for example, by Berner et al. (2015) and Thompson et al. (2021). Having this SPG in ICON allows to reproduce these studies and investigate whether ICON responds similarly to stochastic parameter perturbations. The advantage of a spectral transform ansatz is that it is possible to specify arbitrary horizontal length scales without the need for any MPI communication between processors. An autoregressive process is used to advance the spectral pattern in time.

Being based on spectral transformations, the SPG is more difficult to implement in ICON than in IFS because ICON does not yet provide the necessary and efficient spectral transform methods, simple because ICON is a finite volume model. This is probably the main reason why other stochastic pattern generators have been implemented and used in ICON. The spectral SPG does not need full spectral transforms, though, but only the transformation from spectral space into grid point space (inverse or back transform). This can be done by a direct evaluation of spherical harmonics resp. Fourier modes. On DWD's NEC we can make use of the Advanced Scientific Library (ASL) that provides vectorized codes to do that. For other HPCs the spherical harmonics functions from SHTOOLS are provided as part of this MR.  For the necessary random number generator, we can again use the ASL random number tools or the built-in random number generator.

The scheme has two main subroutines which are both called from `nwp_nh_interface`:

`stochastic_pattern_step`: advances the spectral coefficients in time with the AR1 autoregressive process

`stochastic_pattern_generator`: calculates the stochastic pattern in grid point space from the spectral coefficients

**New namelist parameters:**

`lstochastic_pattern_generator`: logical switch to turn on the SPG.

`spg_fourier_modes`: Use Fourier modes for stochastic pattern generator in LAM configurations. 

`spg_length_scale`: correlation length scale in m.

`spg_time_scale`: time scale of AR1 in s.

`spg_spec_modes`: number of spectral modes in SPG.

`spg_variance`: variance of stochastic pattern in grid-point space.

The spherical harmonics based SPG can indeed be used in ICON LAM simulations but is computationally more expensive than the Fourier version. This can be useful in some experiments, for example for consistency in global and limited-area simulations. Therefore `spg_fourier_modes` is set to TRUE for the LAM mode of ICON.

For efficiency reasons it can be beneficial to run the SPG on a longer time step than the `dtime` of ICON. This is appropriate if the time scale of the random pattern `spg_time_scale` is much longer than `dtime`, which is usually the case. Currently the time step of the spectral SPG is automatically set to a multiple of `dtime` that is 4 times smaller than `spg_time_scale`. For example, a `spg_time_scale` of 30 min with a `dtime` of 120 s leads to an SPG time step of 360 s or 6 min.

Another attempt to improve the efficiency is the selection of the spectral wave modes that are actually calculated. For the spherical harmonics representation the dominant wave mode is calculated from `spg_length_scale` and then the `spg_spec_modes` spherical harmonics modes centered around the dominant mode are used. Note that several spherical harmonics modes are needed to represent a certain length scale on the globe. In LAM mode the first `spg_spec_modes` Fourier modes are used (without an internal check that this number is actually sufficient). In LAM mode, where no Legendre polynomials are need, the efficiency issues are less severe than on the globe.

Multiple patterns with different length or time scales as they are sometimes used in SPPT or SPP of ECMWF have not yet been implemented in this SPG in ICON. This would be a rather straightforward extension of the current module.

**Implementation and performance on DWD's NEC Aurora:**

On the NEC it is important to use vectorized codes for the spherical harmonics and the random number generator. Those are provided by the Advanced Scientific Library (ASL). The ASL can be added to the configure with `-lasl_sequential`. In addition, the preprocessor flag `__ASL__` has to be defined either in the configure or in the module. Using the ASL, the stochastic pattern generator is less expensive than a typical physics scheme like turbulence, microphysics, or deep convection on an R3B6 global grid. Using the unvectorized implementation of SHTOOLS on the NEC comes with a severe performance penalty. Therefore the simulation is aborted, when SHTOOLS is used on the NEC, i.e. when the SPG runs without ASL. If this `CALL finish` is removed, the SHTOOLS code runs on the NEC an gives very similar results as the implementation with ASL.


**References:**

Palmer TN, Buizza R, Doblas-Reyes F, Jung T, Leutbecher M, Shutts G, Steinheimer M, Weisheimer A. 2009. Stochastic Parameterization andModel Uncertainty. Technical Memorandum 598. ECMWF: Reading, UK. 

Ollinaho, P., Lock, S.-J., Leutbecher, M., Bechtold, P., Beljaars, A., Bozzo, A., Forbes, R.M., Haiden, T., Hogan, R.J. and Sandu, I. (2017), Towards process-level representation of model uncertainties: stochastically perturbed parametrizations in the ECMWF ensemble. Q.J.R. Meteorol. Soc., 143: 408-422. https://doi.org/10.1002/qj.2931

Berner, J., Fossell, K. R., Ha, S. Y., Hacker, J. P., & Snyder, C. (2015). Increasing the skill of probabilistic forecasts: Understanding performance improvements from model-error representations. Monthly Weather Review, 143(4), 1295-1320.

Thompson, G., Berner, J., Frediani, M., Otkin, J. A., & Griffin, S. M. (2021). A stochastic parameter perturbation method to represent uncertainty in a microphysics scheme. Monthly Weather Review, 149(5), 1481-1497.


Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Co-authored-by: Thomas Jahns <jahns@dkrz.de>
Approved-by: Christoph Gebhardt <christoph.gebhardt@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>