[nwp] [nwp] Mixed-phase Spectral Bin Microphysics
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1658

Last successful pipeline: https://gitlab.dkrz.de/icon/icon-nwp/-/pipelines/108800

## Feature Description
Following the preliminary warm-sbm version merged with the master on July 2023, we merge now the full mixed-sbm version, which was successfully tested on real cases. This version is based on so called "fast-sbm" version of WRF, which includes 33 advected mass-bins distributions for drops/rain, ice/snow, graupel or hail and ccn. This version includes the processes of nucleation, diffusional growth, collisions, breakup, melting, freezing etc. Importantly, the saturation adjustment is switched off. Moreover, qc and qi which are vertically mixed in turbdiff are diagnosed from the individual bins. 
The plan is to merge mixed-sbm with EMVORADO, as well as to introduce turbulence-dependent collisions enhancement and more sophisticated melting scheme.



Co-authored-by: Ulrich Schaettler <Ulrich.Schaettler@dwd.de>
Co-authored-by: Daniel Rieger <daniel.rieger@dwd.de>
Co-authored-by: Pavel Khain <khainp@ims.gov.il>
Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Co-authored-by: Blahak Ulrich <ulrich.blahak@dwd.de>
Co-authored-by: zil2 Israel TC e-suite <zil2@ac6-100.bullx>
Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>