[nwp] Bugfix: Fix time definition in CAMS climatology preprocessing script
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1784

### Short description

The original CAMS climatology file does not contain a time format that can be read by ICON. The attributes of the time variable are therefore modified in a preprocessing script (modified from a legacy script used for the prior version of the CAMS climatology). The new time format attributes were working in the ICON test experiment, but caused problems in other applications. The time variable is was defined as "months since 2001-1-15 24:00:00", where "24:00" is not a cdf-conform way of defining time (should be 00:00 of the following day).

This merge request modifies the time format slightly (one-liner in preprocessing script) to "months since 2001-1-15 12:00:00" (i.e. middle of the month January). The corresponding CAMS climatology file for the buildbot test should also be updated so that the time attribute matches (although this should make no difference to the buildbot test). The updated file is located here:

/hpc/uwork/mahlgrim/cams/lamtest/icon_cams_clim_lam_test_49r2_DOM01.nc

### 

---

- [x] I have read the contribution guidelines for merge requests (last update: 2025-03-26).
- [x] I have considered updating the [docs.icon-model.org documentation](https://gitlab.dkrz.de/icon/wiki/-/wikis/Documentation).

Co-authored-by: Daniel Reinert <daniel.reinert@dwd.de>
Approved-by: Daniel Rieger <daniel.rieger@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>