Correction of some smaller bugs in observation processing for LHN
https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/1706

LHN is processing radar data. To obtain an observed precipitation rate at any model time steps, an interpolation in time is applied. In this code some bugs were found. It also appeared, that the smoothing of observed and modelled precipitation rate (together with latent heat release of the model) should also be applied to the two arrays of observation weigths (wobs_time and wobs_space) to stay consistent with the smoothed observations. 
Changes took place in:
- time interpolation between obs times
- calculation of time weights (wobs_time)
- initialize lhn_spqual with 1.0
- take into account wobs_space and wobs_time within relaxation to avoid
  areas of observed precipitation set to be zero
- small changes in verification subroutine

In addition to these changes, some namelist variables of LHN were made domain dependent, to allow to change those parameters for nested domains. As the dynamics in domains with higher horizontal resolution might be significantly different to coarser domains, this might be of importance for applying LHN at those scales (i.e. ICON D05 at DWD)

Approved-by: Daniel Leuenberger <daniel.leuenberger@meteoswiss.ch>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>